+++ modules/admin/src/messages/fr/admin.php
@@ -223,4 +223,5 @@ return [
     'model_user_title_mr' => 'Mr.',
     'model_user_title_mrs' => 'Ms.',
     'js_ngrest_toggler_success' => '%field% successfully updated.',
+    'layout_image_filter_selection' => 'Select image Filter:',
 ];